// src/types/core/database.ts
export interface RuleTable {
  rule_id: string;
  rule_name: string;
  rule_type: string;
  rule_details: Json;
  created_by: string;
  status: string;
  created_at: string;
  updated_at: string;
  is_template: boolean;
}

/**
 * Database Rule Insert type
 */
export type RuleInsert = Omit<RuleTable, 'rule_id' | 'created_at' | 'updated_at'> & {
  rule_id?: string;
};